혜진이가 저지른 실수 예원이가 수습함.
이러한 귀차니즘에서 비롯된 오류를 일으킨 사실을 다시 적발 당할 시 사살.
