+++ Assets/Scripts/VolumeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b917e3e3c1f5da45b8333c566faf8ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
