+++ Assets/scriptsforplay/SoundSystem.cs
@@ -9,18 +9,23 @@ public class SoundSystem : MonoBehaviour
 
     public static AudioSource pgMusic;
     public static AudioSource bgm;
+    bool songPlayed=false;
 
     void Start()
     {
         pgMusic=pgMusics[Data.selected_song];
         bgm=bgms[Data.selected_song];
         pgMusic.volume = Data.volumes[0];
-        pgMusic.Play(0);
         bgm.volume = Data.volumes[1];
-        bgm.Play(0);
     }
 
     void Update(){
+        if(NoteManager.loadDelay<=0f&&!songPlayed){
+            pgMusic.Play(0);
+            bgm.Play(0);
+            songPlayed=true;
+        }
+
         bgm.timeSamples=pgMusic.timeSamples;
     }
 }
